refactored reactor tests
